chore: rename Presentation component. improve animation
