+++ src/pages/main/Main.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Header, Canvas, Presentation, Bio } from '../../shared/components';
+import { Header, Canvas, Hero, Bio } from '../../shared/components';
 
 import styles from './Main.css';
 
@@ -7,7 +7,7 @@ const Main = () => (
 	<div className={ styles.root }>
 		<Header />
 		<Canvas isTouchable={ false } />
-		<Presentation />
+		<Hero />
 		<Bio />
 	</div>
 );
