+++ src/shared/components/hero/Hero.css
@@ -3,7 +3,7 @@
 .wrapper {
 	position: sticky;
 	top: 0;
-	z-index: var(--zindex-presentation);
+	z-index: var(--zindex-hero);
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
@@ -20,6 +20,7 @@
 	flex-direction: column;
 	color: var(--color-white-smoke);
 	line-height: 1.7;
+	transition: transform 0.1s linear;
 }
 
 .greetingWrapper {
@@ -46,11 +47,6 @@
 		transform: translateY(0);
 		opacity: 1;
 	}
-}
-
-.greeting,
-.job {
-	transition: transform 0.1s linear;
 }
 
 .name {
