+++ src/shared/components/hero/Hero.js
@@ -2,12 +2,11 @@ import React, { useState, useEffect, useRef } from 'react';
 import classNames from 'class-names';
 import useWindowDimension from '../../hooks/window-dimension/useWindowDimension';
 import { clampValue } from '../../utils';
-import styles from './Presentation.css';
+import styles from './Hero.css';
 
-const Presentation = () => {
+const Hero = () => {
 	const containerRef = useRef();
-	const greetingRef = useRef();
-	const jobRef = useRef();
+	const titleRef = useRef();
 	const [entered, setEntered] = useState(false);
 	const windowDimension = useWindowDimension();
 	const heightThreshold = windowDimension.height;
@@ -23,10 +22,8 @@ const Presentation = () => {
 			containerRef.current.style.position = 'relative';
 		}
 
-		greetingRef.current.style.opacity = 1 - percentage;
-		greetingRef.current.style.transform = `scale(${(percentage * 1) + 1}) translateY(${-percentage*10}rem)`;
-		jobRef.current.style.opacity = 1 - percentage;
-		jobRef.current.style.transform = `scale(${(percentage * 1) + 1})  translateY(${percentage*10}rem)`;
+		titleRef.current.style.opacity = 1 - percentage;
+		titleRef.current.style.transform = `scale(${(percentage * 2) + 1})`;
 	};
 
 
@@ -39,17 +36,17 @@ const Presentation = () => {
 
 	return (
 		<div className={styles.wrapper} ref={containerRef}>
-			<h2 className={styles.title}>
+			<h2 className={styles.title} ref={titleRef}>
 				<div className={classNames(styles.greetingWrapper, entered && styles.entered)}>
-					<div className={styles.greeting} ref={greetingRef}>Hello, I'm</div>
+					Hello, I'm
 				</div>
 				<div className={styles.name}>Gil Domingues</div>
 				<div className={classNames(styles.jobWrapper, entered && styles.entered)}>
-					<div className={styles.job} ref={jobRef}>Software Engineer</div>
+					Software Engineer
 				</div>
 			</h2>
 		</div>
 	);
 };
 
-export default Presentation;
+export default Hero;
