+++ src/shared/components/hero/index.js
@@ -1,0 +1,1 @@
+export { default } from './Hero';
