+++ src/shared/components/index.js
@@ -1,4 +1,4 @@
 export { default as Canvas } from './canvas';
 export { default as Header } from './header';
-export { default as Presentation } from './presentation';
+export { default as Hero } from './hero';
 export { default as Bio } from './bio';
