--- src/shared/components/presentation/index.js
@@ -1,1 +1,0 @@
-export { default } from './Presentation';
