+++ src/shared/styles/variables.css
@@ -10,5 +10,5 @@
 
 	--zindex-header: 1000;
 	--zindex-canvas: 999;
-	--zindex-presentation: 100;
+	--zindex-hero: 100;
 }
